// src/chat/ChatEvent.enum.ts
export enum MessageEvent {
    ADMIN_MESSAGE = "ADMIN_MESSAGE",
    USER_MESSAGE = "USER_MESSAGE"
}

export enum AdminMessageEvent {
    NEW_MEMBER_JOINED = "NEW_MEMBER_JOINED",
    JOINED_SUCCESSFULLY = "JOINED_SUCCESSFULLY",
    JOIN_REJECTED = "JOIN_REJECTED",
    LEAVED_SUCCESSFULLY = "LEAVED_SUCCESSFULLY"
}